requirements.txt
numpy==1.25.2
pillow==10.0.0
pandas==2.1.0
scipy == 1.14.0
scikit-image==0.21.0
opencv-python==4.8.0.76
einops==0.6.1
torch==2.0.1
torchvision==0.15.2
tomli==2.0.1
pytorch-lightning==2.0.8
matplotlib==3.7.2
scikit-learn==1.3.1
umap-learn==0.5.5
seaborn==0.13.1
timm==1.0.9
transformers == 4.44.2
typing